Stepper motor controller
In this project, a stepper motor is to act as a moving car. when the car
has a passenger, a red led is lit, and when it doesn't, a green led is
lit. the passenger is charged 20 dollars for every 5 revolutions of the
motor, displayed to an lcd, the time thet the car moves without a
passenger is also displayed, for every 10 revolutions, 2000 dollars is
charged as fuel